#1 - add-clientApi
